requirements.txt
fastmcp==2.14.1
yfinance
alpaca-trade-api-python
python-dotenv==1.2.1
requests==2.32.5
pandas==2.3.3
pandas_ta
ta==0.11.0
tweepy==4.16.0
praw==7.8.1
newsapi-python==0.2.7
RestrictedPython==8.1
websockets==15.0.1
feedparser==6.0.11
fastapi==0.115.5
uvicorn==0.32.1